+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### NEXT
 * Introduce support for ASN.1 REAL
+* Add built-in ASN.1 ENUMERATED support
 * Rename `ObjectIdentifier.parse` -> `ObjectIdentifier.decodeFromAsn1ContentBytes` in accordance with other similar functions
 * Add dedicated Android targets (SDK 30 / JDK 1.8) to all modules
 * Fix internal deprecations